refactor: List와 Grid Cell의 상위 클래스인 MarketCollectionViewCell 생성
- 중복된 뷰와 메서드를 상위 클래스에서 구현한 후 상속을 통해 추상화
